Refactoring - part 16
