+++ src/main/kotlin/kscript/app/exception/InvalidDependencyLocatorException.kt
@@ -1,3 +1,3 @@
-package kscript.app.common
+package kscript.app.exception
 
 class InvalidDependencyLocatorException(locator: String) : RuntimeException("Invalid dependency locator: '$locator'")
